+++ packages/build-ui/src/pages/agits/_id/management/artworks/components/mod.rs
@@ -1,0 +1,3 @@
+mod artwork_table;
+
+pub use artwork_table::ArtworkTable;
