+++ packages/build-ui/src/pages/agits/_id/management/collections/components/mod.rs
@@ -1,9 +1,12 @@
-pub mod collection_name_input_modal;
-pub mod new_collection_modal;
-pub mod success_modal;
-pub mod transfer_confirmation_modal;
+mod collection_name_input_modal; 
+mod new_collection_modal;
+mod success_modal;
+mod transfer_confirmation_modal;
+mod collections_table;
+
 
 pub use collection_name_input_modal::CollectionNameInputModal;
 pub use new_collection_modal::NewCollectionModal;
 pub use success_modal::SuccessModal;
 pub use transfer_confirmation_modal::TransferConfirmationModal;
+pub use collections_table::CollectionTable;
