+++ packages/build-ui/src/pages/agits/_id/management/collectors/components/mod.rs
@@ -1,0 +1,2 @@
+mod collectors_table;
+pub use collectors_table::CollectorsTable;
