+++ packages/build-ui/src/pages/agits/_id/management/collectors/mod.rs
@@ -2,6 +2,7 @@ mod _id;
 mod controllers;
 mod i18n;
 mod page;
+mod components;
 
 pub use _id::CollectorDetailPage;
 pub use page::CollectorsPage;
